# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pulse Messenger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /calls:
    get:
      operationId: getCallHistory
      tags: [calls]
      summary: Get call history
      responses:
        "200":
          description: Call history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Call"
    post:
      operationId: initiateCall
      tags: [calls]
      summary: Initiate a call
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateCallBody"
      responses:
        "201":
          description: Call initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
  /calls/{callId}:
    get:
      operationId: getCallById
      tags: [calls]
      parameters:
        - name: callId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
    put:
      operationId: updateCallStatus
      tags: [calls]
      parameters:
        - name: callId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCallBody"
      responses:
        "200":
          description: Updated call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        avatarColor:
          type: string
        status:
          type: string
          enum: [online, offline, away]
        statusText:
          type: string
          nullable: true
        lastSeen:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        isBot:
          type: boolean
        isVerified:
          type: boolean
        isAdmin:
          type: boolean
        showOnlineStatus:
          type: boolean
        readReceiptsEnabled:
          type: boolean
        totpEnabled:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - displayName
        - avatarColor
        - status
        - isBot
        - isVerified
        - isAdmin
        - createdAt
...
    Call:
      type: object
      properties:
        id:
          type: integer
        chatId:
          type: integer
          nullable: true
        callerId:
          type: integer
        calleeId:
          type: integer
          nullable: true
        type:
          type: string
          enum: [audio, video, group]
        status:
          type: string
          enum: [ringing, active, ended, missed, declined]
        startedAt:
          type: string
          nullable: true
        endedAt:
          type: string
          nullable: true
        durationSeconds:
          type: integer
          nullable: true
        caller:
          $ref: "#/components/schemas/User"
          nullable: true
        callee:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - callerId
        - type
        - status
        - createdAt
    InitiateCallBody:
      type: object
      properties:
        calleeId:
          type: integer
        chatId:
          type: integer
        type:
          type: string
          enum: [audio, video, group]
      required:
        - type
    UpdateCallBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, ended, declined, missed]
      required:
        - status
...